requirements.txt
python-socketio
requests
aiohttp
scikit-learn
numpy
sortedcontainers
